+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Deprecate `deviceSessionId` and `fraudMerchantId` in `CreditCardRequest`, `CustomerRequest`, `PaymentMethodRequest`, and `TransactionRequest` classes
 * Fix issue where risk data was not included in `CustomerRequest` (fixes #85)
 * Add `getInstallmentCount()` method to `dispute.getTransaction()` on dispute webhooks
+* Update `lookup` method on `ThreeDSecureGateway` to return a `Result` object
 
 ## 3.1.0
 * Add `RISK_THRESHOLD` to `GatewayRejectionReason`